requirements.txt
astroid==2.2.5
autopep8==1.4.4
boto3==1.9.215
botocore==1.12.215
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.3
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-storages==1.7.1
docutils==0.15.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
pycryptodome==3.8.2
pylint==2.3.1
PyPDF2==1.26.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.3
wrapt==1.11.2
xhtml2pdf==0.2.3
